[Add] Projects 파이어베이스 데이터베이스와 연동 완료
